refactor: organizes styles
