+++ data/articles/2021-10-09-what-is-getstaticprops.mdx
@@ -109,7 +109,7 @@ The `context` parameters consists of the following keys:
 | `preview`       | Depends on the [preview mode](https://nextjs.org/docs/advanced-features/preview-mode)                                               | `true` or `undefined`   |
 | `previewData`   | Data set by [`setPreviewData`](https://nextjs.org/docs/advanced-features/preview-mode#step-1-create-and-access-a-preview-api-route) | `en-US`                 |
 | `locale`        | Current locale of your [internationalized Next.js application](https://nextjs.org/docs/advanced-features/i18n-routing)              | `en-US`                 |
-| `locales`       | All the locales of your [internationalized Next.js application](https://nextjs.org/docs/advanced-features/i18n-routing              | `en-US` or `fr`         |
+| `locales`       | All the locales of your [internationalized Next.js application](https://nextjs.org/docs/advanced-features/i18n-routing)             | `en-US` or `fr`         |
 | `defaultLocale` | Default locale of your [internationalized Next.js application](https://nextjs.org/docs/advanced-features/i18n-routing)              | `en-US`                 |
 
 ## Return value of getStaticProps
